# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  daemon.ExecutionStatePageData:
    properties:
      execution:
        $ref: '#/definitions/github_com_thand-io_agent_internal_models.WorkflowExecutionInfo'
      workflow:
        $ref: '#/definitions/model.Workflow'
    type: object
...
  models.ElevateLLMRequest:
    properties:
      reason:
        type: string
    type: object
  models.ElevateRequest:
    properties:
      authenticator:
        description: Which provider to use for authentication
        type: string
      duration:
        description: Duration in ISO 8601 format
        type: string
      identities:
        description: Optional identities to elevate, if empty the requesting user
          is used
        items:
          type: string
        type: array
      providers:
        description: A role can be applied to multiple providers
        items:
          type: string
        type: array
      reason:
        type: string
      role:
        $ref: '#/definitions/github_com_thand-io_agent_internal_models.Role'
      session:
        $ref: '#/definitions/github_com_thand-io_agent_internal_models.LocalSession'
      tenants:
        description: Optional tenant IDs for multi-account providers
        items:
          type: string
        type: array
      workflow:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@thand.io
    name: Thand Support
    url: https://thand.io/support
  description: Thand Agent API for managing access, roles, workflows, and providers
  license:
    name: BSL
    url: https://mariadb.com/bsl11/
  termsOfService: https://thand.io/terms
  title: Thand Agent API
  version: "1.0"
paths:
...
  /auth:
    get:
      consumes:
      - application/json
      description: Display the authentication page with available providers
      parameters:
      - description: Callback URL
        in: query
        name: callback
        type: string
      - description: Provider name for direct authentication
        in: query
        name: provider
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: Authentication page
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      summary: Authentication page
      tags:
      - auth
  /auth/callback/{provider}:
    get:
      consumes:
      - application/json
      description: Handle the OAuth2 callback from the provider
      parameters:
      - description: Provider name
        in: path
        name: provider
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      - description: OAuth code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Authentication successful
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      summary: OAuth2 authentication callback
      tags:
      - auth
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Handle the SAML POST callback from the provider
      parameters:
      - description: Provider name
        in: path
        name: provider
        required: true
        type: string
      - description: SAML RelayState (SP-initiated)
        in: formData
        name: RelayState
        type: string
      - description: SAML Response
        in: formData
        name: SAMLResponse
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Authentication successful
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      summary: SAML authentication callback
      tags:
      - auth
  /auth/logout:
    get:
      consumes:
      - application/json
      description: Clear the user session and logout
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            additionalProperties: true
            type: object
        "307":
          description: Redirect to home page
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Logout
      tags:
      - auth
  /auth/logout/{provider}:
    get:
      consumes:
      - application/json
      description: Clear the user session and logout
      parameters:
      - description: Provider name
        in: path
        name: provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            additionalProperties: true
            type: object
        "307":
          description: Redirect to home page
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Logout
      tags:
      - auth
  /auth/request/{provider}:
    get:
      consumes:
      - application/json
      description: Start the OAuth2 authentication flow for a provider
      parameters:
      - description: Provider name
        in: path
        name: provider
        required: true
        type: string
      - description: Callback URL
        in: query
        name: callback
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirect to provider authentication
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Provider not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Initiate authentication
      tags:
      - auth
  /elevate:
    get:
      consumes:
      - application/json
      description: Request elevation to a specific role with static parameters
      parameters:
      - description: Role name
        in: query
        name: role
        required: true
        type: string
      - description: Provider name
        in: query
        name: provider
        required: true
        type: string
      - description: Reason for elevation
        in: query
        name: reason
        required: true
        type: string
      - description: Duration of elevation
        in: query
        name: duration
        type: string
      - description: Workflow name
        in: query
        name: workflow
        type: string
      - description: Identity filter
        in: query
        name: identities
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Elevation request submitted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      summary: Request role elevation
      tags:
      - elevate
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      - multipart/form-data
      description: Submit an elevation request with dynamic or static parameters
      parameters:
      - description: Elevation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ElevateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Elevation request submitted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      summary: Submit elevation request
      tags:
      - elevate
  /elevate/llm:
    get:
      consumes:
      - application/json
      description: Request elevation using natural language reasoning with LLM
      parameters:
      - description: Natural language reason for elevation
        in: query
        name: reason
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: LLM response with suggested role
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      summary: LLM-based elevation
      tags:
      - elevate
    post:
      consumes:
      - application/json
      description: Request elevation using natural language with POST request
      parameters:
      - description: LLM elevation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ElevateLLMRequest'
      produces:
      - application/json
      responses:
        "200":
          description: LLM response with suggested role
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      summary: LLM-based elevation (POST)
      tags:
      - elevate
  /elevate/resume:
    get:
      consumes:
      - application/json
      description: Resume a paused or interrupted elevation workflow
      parameters:
      - description: Workflow state token
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirect to next workflow step
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      summary: Resume elevation workflow
      tags:
      - elevate
    post:
      consumes:
      - application/json
      description: Resume a paused elevation workflow with POST data
      parameters:
      - description: Resume request data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "307":
          description: Redirect to next workflow step
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      summary: Resume elevation workflow (POST)
      tags:
      - elevate
  /execution:
    post:
      consumes:
      - application/json
      description: Create and start a new workflow execution
      parameters:
      - description: Workflow creation request
        in: body
        name: workflow
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Workflow created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create workflow execution
      tags:
      - executions
  /execution/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific workflow execution
      parameters:
      - description: Workflow execution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Workflow execution details
          schema:
            $ref: '#/definitions/daemon.ExecutionStatePageData'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get workflow execution
      tags:
      - executions
  /execution/{id}/cancel:
    get:
      consumes:
      - application/json
      description: Gracefully cancel a running workflow execution
      parameters:
      - description: Workflow execution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Workflow cancelled
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Workflow not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel workflow execution
      tags:
      - executions
  /execution/{id}/signal:
    get:
      consumes:
      - application/json
      description: Send a signal event to a running workflow execution
      parameters:
      - description: Workflow execution ID
        in: path
        name: id
        required: true
        type: string
      - description: Encoded signal data
        in: query
        name: input
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Signal sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Signal workflow execution
      tags:
      - executions
...
  /provider/{provider}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific provider
      parameters:
      - description: Provider name
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Provider details
          schema:
            $ref: '#/definitions/models.ProviderResponse'
        "404":
          description: Provider not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get provider by name
      tags:
      - providers
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"